+++ branch-chain-aggregator/src/transaction/leap.rs
@@ -1,0 +1,10 @@
+use crate::error::Error;
+use crate::Aggregator;
+
+use ckb_types::H256;
+
+impl Aggregator {
+    pub(crate) fn create_leap_tx(&self) -> Result<H256, Error> {
+        todo!()
+    }
+}
